refactor: split MySQL DSN option file and SQL generation
